refactor(deleteButton): function as composable
