+++ src/components/Layout/Views/UserProfileView.vue
@@ -57,7 +57,6 @@
       </Boton>
     </div>
   </div>
-
   <div
     style="justify-content: center; display: flex;"
     @click="deleteAccountMessage">
@@ -76,42 +75,11 @@
 
 
 <script setup lang="ts">
-import Swal from 'sweetalert2';
-import { useI18n } from 'vue-i18n';
 import { useRouter } from 'vue-router/auto';
 import { auth } from '@/store/auth';
+import { deleteAccountMessage } from '@/composables/use-delete-account-alert';
 
-const { t } = useI18n();
 const router = useRouter();
-
-/**
- * Esta función se encarga de preguntar al usuario si quiere borrar su cuenta con un mensaje
- */
-async function deleteAccountMessage() : Promise<void> {
-  await Swal.fire({
-    title: t('deleteAccountMessageTitle'),
-    text: t('deleteAccountMessageText'),
-    icon: 'warning',
-    showCancelButton: true,
-    confirmButtonColor: '#910e38',
-    cancelButtonColor: '#0e4781',
-    confirmButtonText: t('deleteAccountMessageConfirm'),
-    cancelButtonText: t('deleteAccountMessageCancel')
-  }).then(async (res) => {
-    if (res.isConfirmed) {
-      await Swal.fire({
-        title: t('deleteAccountMessageConfirmedTitle'),
-        text: t('deleteAccountMessageConfirmedText'),
-        icon: 'success'
-      });
-      auth.logout();
-
-      return 'Account deleted';
-    } else {
-      return 'Account deletion cancelled';
-    }
-  });
-}
 </script>
 
 <style scoped>
